Extract and use MovieClickListener.
